fix problem in the count of items, and more modular and scalable method to create new tabs in pokemons.cs
Created items panel in a separated file as user component.
